refactor: extract RichTextEditor component from existing components
